address PR feedbacks
