+++ src/language/languageServerController.ts
@@ -38,7 +38,9 @@ export default class LanguageServerController {
   _currentConnectionId: string | null = null;
   _currentConnectionString?: string;
   _currentConnectionOptions?: MongoClientOptions;
-  _consoleOutputChannel?: vscode.OutputChannel;
+
+  _consoleOutputChannel =
+    vscode.window.createOutputChannel('Playground output');
 
   constructor(context: ExtensionContext) {
     this._context = context;
@@ -158,12 +160,12 @@ export default class LanguageServerController {
       ServerCommands.SHOW_CONSOLE_OUTPUT,
       (outputs) => {
         for (const line of outputs) {
-          this._consoleOutputChannel?.appendLine(
+          this._consoleOutputChannel.appendLine(
             typeof line === 'string' ? line : util.inspect(line)
           );
         }
 
-        this._consoleOutputChannel?.show(true);
+        this._consoleOutputChannel.show(true);
       }
     );
   }
@@ -187,6 +189,8 @@ export default class LanguageServerController {
       inputLength: playgroundExecuteParameters.codeToEvaluate.length,
     });
     this._isExecutingInProgress = true;
+
+    this._consoleOutputChannel.clear();
 
     // Instantiate a new CancellationTokenSource object
     // that generates a cancellation token for each run of a playground.
