+++ src/language/serverCommands.ts
@@ -10,6 +10,7 @@ export enum ServerCommands {
   CLEAR_CACHED_COMPLETIONS = 'CLEAR_CACHED_COMPLETIONS',
   MONGODB_SERVICE_CREATED = 'MONGODB_SERVICE_CREATED',
   INITIALIZE_MONGODB_SERVICE = 'INITIALIZE_MONGODB_SERVICE',
+  CODE_EXECUTION_RESULT = 'CODE_EXECUTION_RESULT',
   SHOW_CONSOLE_OUTPUT = 'SHOW_CONSOLE_OUTPUT',
 }
